backend/requirements.txt
fastapi
uvicorn
PyPDF2
pdf2image
pillow
transformers
torch
olmocr
python-multipart
